requirements.txt
PyYAML
chess
